api/requirements.txt
-i https://pypi.org/simple
asgiref==3.2.3
django-cors-headers==3.2.1
django==3.0.2
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
